Use go's context.Context
This makes the existing code a bit more idiomatic by using the builtin
context.Context type in place of ExecutionContext.  Timeouts are handled
through context.WithTimeout, context.DeadlineExceeded is used in place
of ErrExecutionTimeout, and context.Value is used to store the
environment variable mapping.

This also allows for cancellation functions to be passed down as a
future enhancement.

It also adds a guarantee around consistency - the command is started on
Invoke, and the timeout captures the entire run of the command.  (In
practice, the time spent executing before the timeout took effect was
pretty negligible, but nice to have the guarantee.)